+++ coday-challenge/src/main/java/com/nice/avishkar/Constants.java
@@ -1,0 +1,13 @@
+package com.nice.avishkar;
+
+public class Constants {
+
+    public static final String UTF_8 = "UTF-8";
+    public static final String VOTER = "Voter";
+    public static final String CONSTITUENCY = "Constituency";
+    public static final String POLLING_STATION = "PollingStation";
+    public static final String CANDIDATE = "Candidate";
+    public static final String NOTA = "NOTA";
+    public static final String NO_WINNER = "NO_WINNER";
+
+}
