+++ coday-challenge/src/main/java/com/nice/avishkar/ConstituencyResult.java
@@ -5,6 +5,9 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
 
+import static com.nice.avishkar.Constants.NOTA;
+import static com.nice.avishkar.Constants.NO_WINNER;
+
 public class ConstituencyResult {
 
 	private String winnerName;
@@ -49,9 +52,9 @@ public class ConstituencyResult {
 
 	public void finalUpdate() {
 		Comparator<CandidateVotes> comparator = (vote1,vote2) -> {
-			if (vote1.getCandidateName().equals("NOTA")) {
+			if (vote1.getCandidateName().equals(NOTA)) {
 				return 1;
-			} else if (vote2.getCandidateName().equals("NOTA")) {
+			} else if (vote2.getCandidateName().equals(NOTA)) {
 				return -1;
 			} else if (vote2.getVotes() > vote1.getVotes()) {
 				return 1;
@@ -61,8 +64,9 @@ public class ConstituencyResult {
 				return vote1.getCandidateName().compareTo(vote2.getCandidateName());
 			}
 		};
-		Collections.sort(this.candidateList, comparator);
-		this.winnerName = candidateList.get(0).getVotes() == candidateList.get(1).getVotes() ? "NO_WINNER" : candidateList.get(0).getCandidateName();
+
+		Collections.sort(this.candidateList,comparator);
+		this.winnerName = candidateList.get(0).getVotes() == candidateList.get(1).getVotes() ? NO_WINNER : candidateList.get(0).getCandidateName();
 	}
 	@Override
 	public String toString() {
