+++ coday-challenge/src/main/java/com/nice/avishkar/Solution.java
@@ -5,7 +5,9 @@ import com.nice.avishkar.dao.CandidatesDaoImpl;
 import com.nice.avishkar.dao.VotersDao;
 import com.nice.avishkar.dao.VotersDaoImpl;
 import com.nice.avishkar.service.CandidateService;
+import com.nice.avishkar.service.CandidateServiceImpl;
 import com.nice.avishkar.service.VoteCountingService;
+import com.nice.avishkar.service.VoteCountingServiceImpl;
 
 import java.nio.file.Path;
 import java.util.Map;
@@ -14,12 +16,19 @@ public class Solution {
 
 	public ElectionResult execute(Path candidateFile, Path votingFile) {
 		ElectionResult resultData = new ElectionResult();
+		//Creating Dao's
 		CandidatesDao candidatesDao = new CandidatesDaoImpl();
 		VotersDao votersDao = new VotersDaoImpl();
-		CandidateService candidateService = new CandidateService(candidatesDao);
+
+		//Creating Services and Injecting Dao
+		CandidateService candidateService = new CandidateServiceImpl(candidatesDao);
+		VoteCountingService voteCountingService = new VoteCountingServiceImpl(votersDao);
+
+		//Getting Constituency To Candidate Data Mapped to avoid Unmapped Candidates
 		Map<String, ConstituencyResult> constituencyToCandidateMap = candidateService.getConstituencyToCandidateMap(candidateFile);
-		VoteCountingService voteCountingService = new VoteCountingService(votersDao);
+
 		Map<String, ConstituencyResult> result = voteCountingService.getResult(votingFile, constituencyToCandidateMap);
+
 		resultData.setResultData(result);
 		return resultData;
 	}
