+++ coday-challenge/src/main/java/com/nice/avishkar/entities/CandidateConstituency.java
@@ -3,10 +3,14 @@ package com.nice.avishkar.entities;
 import com.univocity.parsers.annotations.Parsed;
 import lombok.Data;
 
+import static com.nice.avishkar.Constants.CONSTITUENCY;
+import static com.nice.avishkar.Constants.CANDIDATE;
+
 @Data
 public class CandidateConstituency<T> {
-    @Parsed(field = "Constituency")
+
+    @Parsed(field = CONSTITUENCY)
     String Constituency;
-    @Parsed(field = "Candidate")
+    @Parsed(field = CANDIDATE)
     String Candidate;
 }
